Fix unclear jira plugin error message (#45)
Fix https://github.com/abcxyz/jvs/issues/325

The original issue is that errors are added to both `err` and
`ValidateJustificationResponse.Error`. To handle this, we categorize the
errors as follows:
- If an error originates from the Jira endpoint or from result/input
validation, it is considered a Validation Error. In this case, we return
a `ValidateJustificationResponse` with an error message. Otherwise, the
error is considered an internal error and is returned as such.
- Both types of errors are handled at the top level. Error details are
logged, and only standard instructions are returned to the user.
([Validation error](https://screenshot.googleplex.com/3gS5cPGwmQynPSx)
and [internal error](https://screenshot.googleplex.com/KRFeXAD7maK7BGN))
- Additionally the Go plugin only supports the Go standard log library
and the Hashicorp logger
[doc](https://github.com/hashicorp/go-plugin#features). From my
experiments, it does not work with the `github.com/abcxyz/pkg/logging`
package. So, I use the Go standard log library in this in stead. Logs
are then marked in debug level by the plugin ([Validation error
logs](https://screenshot.googleplex.com/6UkYx4NpSivDWwJ) and [Internal
error logs](https://screenshot.googleplex.com/4hyshunYJFupv7k))

---------

Signed-off-by: jiahaluo <116128330+jiahaluo@users.noreply.github.com>
Co-authored-by: Chen Shou <cshou@google.com>